+++ CHANGELOG.md
@@ -70,6 +70,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `simple` quest items now resolve variables one time on reload to support `constant` variables
     - `owner:%player%` was changed to `owner:` to allow constant pre-parsing
 - `math` variable now gives a better exception message and does not return 0 instead
+- `folder` event now use the `unit` argument to define the time unit for ticks, seconds and minutes
 ### Deprecated
 ### Removed
 - undocumented prefix feature in conversation